requirements.txt
halo==0.0.26
crayons==0.2.0
Click==7.0
requests==2.22.0
GitPython==2.1.11
# ontology-python-sdk==2.0.6